implement lisgan model in pipeline
