+++ config.yaml
@@ -1,0 +1,105 @@
+CADA_VAE:
+  CROSS_RECONSTRUCTION: true
+  DISTRIBUTION_ALLIGNMENT: true
+  DROPOUT_RATE: 0
+  HIDDEN_SIZE:
+    DECODER:
+      CLS_ATTR:
+      - 665
+      IMG:
+      - 1660
+    ENCODER:
+      CLS_ATTR:
+      - 1450
+      IMG:
+      - 1560
+  LATENT_SIZE: 64
+  NORM_TYPE: l1
+  USE_BN: false
+  WARMUP:
+    BETA:
+      END_EPOCH: 93
+      FACTOR: 0.25
+      START_EPOCH: 0
+    CROSS_RECONSTRUCTION:
+      END_EPOCH: 75
+      FACTOR: 2.37
+      START_EPOCH: 21
+    DISTANCE:
+      END_EPOCH: 22
+      FACTOR: 8.13
+      START_EPOCH: 6
+CLS:
+  BATCH_SIZE: 128
+  EPOCH: 20
+  RATIO: 0.2
+  SOLVER:
+    AMSGRAD: false
+    BASE_LR: 0.001
+    BETAS:
+    - 0.9
+    - 0.999
+    DAMPENING: 0.0
+    GAMMA: 0.1
+    LR_POLICY: ''
+    MOMENTUM: 0.9
+    NESTEROV: true
+    OPTIMIZING_METHOD: adam
+    WEIGHT_DECAY: 0.0001
+CUDNN_DETERMINISTIC: false
+DATA:
+  FEAT_EMB:
+    DIM:
+      CLS_ATTR: 312
+      IMG: 2048
+    PATH: data/CUB/resnet101/
+  NUM_CLASSES: 200
+  ORIG:
+    PATH: ''
+  TRAIN_CLS_NUM: 150
+DATA_LOADER:
+  DROP_LAST: true
+  NUM_WORKERS: 8
+  PIN_MEMORY: true
+DEVICE: cpu
+GENERALIZED: true
+LISGAN:
+  BETA: 10
+  CLASSIFIER_NUM_EPOCH: 100
+  CLS_NUM_SAMPLES: 300
+  CLS_WEIGHT: 0.01
+  HIDDEN_SIZE:
+    DISCRIMINATOR: 1024
+    GENERATOR: 4096
+  NOISE_SIZE: 312
+  NUM_SYNTH_SAMPLES: 20
+  N_CLUSTER: 3
+  REG1_COEF: 0.01
+  REG2_COEF: 0.001
+  TRAIN_DISCRIMINATOR_NUM_ITER: 5
+LOG_MODEL_INFO: true
+LOG_PERIOD: 10
+NUM_SHOTS: 0
+OUTPUT_DIR: .
+RNG_SEED: -1
+VERBOSE: 2
+ZSL:
+  BATCH_SIZE: 128
+  EPOCH: 1
+  SAMPLES_PER_CLASS:
+    CLS_ATTR: 400
+    IMG: 200
+  SOLVER:
+    AMSGRAD: false
+    BASE_LR: 0.0001
+    BETAS:
+    - 0.5
+    - 0.999
+    DAMPENING: 0.0
+    GAMMA: 0.1
+    LR_POLICY: ''
+    MOMENTUM: 0.9
+    NESTEROV: true
+    OPTIMIZING_METHOD: adam
+    WEIGHT_DECAY: 0.0001
+ZSL_MODEL_NAME: LisGAN
