+++ zeroshoteval/proc/zeroshotnet/__init__.py
@@ -3,3 +3,4 @@ from .build import ZSL_MODEL_REGISTRY, build_zsl
 
 # import all zsl procedures sothey can be registered
 from .cada_vae.cada_vae_train import CADA_VAE_train_procedure
+from .lisgan.lisgan_train import LisGAN_train_procedure
